Workaround buggy PHP warning
The recent change in https://github.com/symfony/process/commit/709e15e7a37cb7ed6199548dc70dc33168e6cb2d causes an exception to be thrown if PHP is configured with a non-empty `open_basedir`.

As soon as this change was rolled-out to our various environments, it began causing show-stopping 500 errors in our Laravel apps. Unlike Symfony, Laravel throws `\ErrorException` even for warnings, and even in production mode.

I'm not familiar with Symfony's testing regimen, but given the consequences of this type of regression outside of Symfony apps, it may be worth running tests with a non-empty `open_basedir` value in the future.

My point is that nobody seemed to notice this issue because it doesn't happen when `open_basedir` is empty (null, empty string, etc.). One may confirm this by defining `open_basedir` as follows, and then loading this class:

```
ini_set('open_basedir', ''); // Warning does NOT occur
ini_set('open_basedir', '/any/path'); //Warning occurs
```

I fixed the same type of issue with the `ExecutableFinder` class in https://github.com/symfony/symfony/pull/16182 (as @nicolas-grekas may recall) , so it seems that the same change is appropriate here.