[Process] Fixed inconsistent test
Sometimes the process no longer appears to be running when the signal is sent which causes a LogicException to be thrown.
This doesn't appear to be consistent and I can reproduce it randomly on my local machine. To avoid having tests fail at random
I decided that it's better to send the signal only if the process is still marked as running.